Fixed mantis #0002588: "Video memory left" is reported as negative when RAM is >= 4GB

+++ mediaportal/xbmc/MediaPortal.cs
@@ -820,7 +820,7 @@ public class MediaPortalApp : D3DApp, IRender
     {
       m_bShowStatsPrevious = false;
     }
-    
+
     if (m_bShowStats)
     {
       GetStats();
@@ -1048,7 +1048,7 @@ public class MediaPortalApp : D3DApp, IRender
 
   // we only reopen the DB connections if the DB path is remote.      
   // since local db's have no problems.
-  private void ReOpenDBs ()
+  private void ReOpenDBs()
   {
     string dbPath = FolderSettings.DatabaseName;
     bool isRemotePath = (string.IsNullOrEmpty(dbPath) || PathIsNetworkPath(dbPath));
@@ -1057,7 +1057,7 @@ public class MediaPortalApp : D3DApp, IRender
       Log.Info("Main: reopen FolderDatabase3 sqllite database.");
       FolderSettings.ReOpen();
     }
-    dbPath = MediaPortal.Picture.Database.PictureDatabase.DatabaseName; 
+    dbPath = MediaPortal.Picture.Database.PictureDatabase.DatabaseName;
     isRemotePath = (string.IsNullOrEmpty(dbPath) || PathIsNetworkPath(dbPath));
     if (isRemotePath)
     {
@@ -1076,7 +1076,7 @@ public class MediaPortalApp : D3DApp, IRender
     isRemotePath = (string.IsNullOrEmpty(dbPath) || PathIsNetworkPath(dbPath));
     if (isRemotePath)
     {
-      Log.Info("Main: reopen MusicDatabase.db3 sqllite database.");      
+      Log.Info("Main: reopen MusicDatabase.db3 sqllite database.");
       MediaPortal.Music.Database.MusicDatabase.ReOpen();
     }
   }
@@ -1101,7 +1101,7 @@ public class MediaPortalApp : D3DApp, IRender
       MediaPortal.Picture.Database.PictureDatabase.Dispose();
     }
 
-    dbPath = MediaPortal.Video.Database.VideoDatabase.DatabaseName; 
+    dbPath = MediaPortal.Video.Database.VideoDatabase.DatabaseName;
     isRemotePath = (!string.IsNullOrEmpty(dbPath) && PathIsNetworkPath(dbPath));
     if (isRemotePath)
     {
@@ -1115,10 +1115,10 @@ public class MediaPortalApp : D3DApp, IRender
     {
       Log.Info("Main: disposing MusicDatabase db3 sqllite database.");
       MediaPortal.Music.Database.MusicDatabase.Dispose();
-    }   
+    }
   }
 
-  private bool Currentmodulefullscreen ()
+  private bool Currentmodulefullscreen()
   {
     bool currentmodulefullscreen = (GUIWindowManager.ActiveWindow == (int)GUIWindow.Window.WINDOW_TVFULLSCREEN ||
                                       GUIWindowManager.ActiveWindow == (int)GUIWindow.Window.WINDOW_FULLSCREEN_MUSIC ||
@@ -1150,8 +1150,8 @@ public class MediaPortalApp : D3DApp, IRender
         while (GUIGraphicsContext.IsPlaying)
         {
           Thread.Sleep(100);
-        }        
-      }      
+        }
+      }
 
       SaveLastActiveModule();
 
@@ -1497,7 +1497,7 @@ public class MediaPortalApp : D3DApp, IRender
     if (showLastActiveModule)
     {
       using (Settings xmlreader = new MPSettings())
-      {        
+      {
         // if MP was closed/hibernated by the use of remote control, we have to retrieve the fullscreen state in an alternative manner.
 
         //if currentmoduleid is TV fullscreen, then set currentmoduleid to tvhome.
@@ -1517,7 +1517,7 @@ public class MediaPortalApp : D3DApp, IRender
         {
           currentmoduleid = Convert.ToString(GUIWindowManager.GetPreviousActiveWindow());
         }
-        
+
 
         if (!currentmodulefullscreen && currentmodulefullscreenstate == "True")
         {
@@ -1566,7 +1566,7 @@ public class MediaPortalApp : D3DApp, IRender
           xmlreader.SetValue(section, "lastfolder", lastFolder);
           Log.Debug("Main: reverting to root folder, pin protected folder was open, SaveLastFolder {0}", lastFolder);
         }
-        
+
         xmlreader.SetValue("general", "lastactivemodule", currentmoduleid);
         xmlreader.SetValueAsBool("general", "lastactivemodulefullscreen", currentmodulefullscreen);
         Log.Debug("Main: SaveLastActiveModule - module {0}", currentmoduleid);
@@ -1579,7 +1579,7 @@ public class MediaPortalApp : D3DApp, IRender
   /// OnExit() Gets called just b4 application stops
   /// </summary>
   protected override void OnExit()
-  {    
+  {
     SaveLastActiveModule();
 
     Log.Info("Main: Exiting");
@@ -1725,7 +1725,7 @@ public class MediaPortalApp : D3DApp, IRender
     {
       Log.Info("Main: DX9 size: {0}x{1}", GUIGraphicsContext.DX9Device.PresentationParameters.BackBufferWidth,
                GUIGraphicsContext.DX9Device.PresentationParameters.BackBufferHeight);
-      Log.Info("Main: Video memory left: {0} MB", GUIGraphicsContext.DX9Device.AvailableTextureMemory / 1048576);
+      Log.Info("Main: Video memory left: {0} MB", (uint)GUIGraphicsContext.DX9Device.AvailableTextureMemory / 1048576);
       // consider the aperture size
     }
     SetupCamera2D();
@@ -1781,69 +1781,69 @@ public class MediaPortalApp : D3DApp, IRender
   /// etc., that don't change during rendering can be set once here to
   /// avoid redundant state setting during Render() or FrameMove().
   /// </summary>
-	protected override void OnDeviceReset(Object sender, EventArgs e)
+  protected override void OnDeviceReset(Object sender, EventArgs e)
   {
-  	// Only perform the device reset if we're not shutting down MediaPortal.
-  	if (GUIGraphicsContext.CurrentState != GUIGraphicsContext.State.STOPPING)
-  	{
-  		Log.Info("Main: OnDeviceReset called");
+    // Only perform the device reset if we're not shutting down MediaPortal.
+    if (GUIGraphicsContext.CurrentState != GUIGraphicsContext.State.STOPPING)
+    {
+      Log.Info("Main: OnDeviceReset called");
 
-  		// Only perform the device reset if we're not shutting down MediaPortal.
-  		if (GUIGraphicsContext.CurrentState != GUIGraphicsContext.State.STOPPING)
-  		{
-  			Log.Info("Main: Resetting DX9 device");
+      // Only perform the device reset if we're not shutting down MediaPortal.
+      if (GUIGraphicsContext.CurrentState != GUIGraphicsContext.State.STOPPING)
+      {
+        Log.Info("Main: Resetting DX9 device");
 
-  			int activeWin = GUIWindowManager.ActiveWindow;
-  			GUIWindowManager.UnRoute();
-  			// avoid that there is an active Window when GUIWindowManager.ActivateWindow(activeWin); is called
-  			Log.Info("Main: UnRoute - done");
+        int activeWin = GUIWindowManager.ActiveWindow;
+        GUIWindowManager.UnRoute();
+        // avoid that there is an active Window when GUIWindowManager.ActivateWindow(activeWin); is called
+        Log.Info("Main: UnRoute - done");
 
-  			GUITextureManager.Dispose();
-  			GUIFontManager.Dispose();
+        GUITextureManager.Dispose();
+        GUIFontManager.Dispose();
 
-  			GUIGraphicsContext.DX9Device.EvictManagedResources();
-  			GUIWaitCursor.Dispose();
-  			if (!m_strSkin.Equals(GUIGraphicsContext.Skin))
-  			{
-  				m_strSkin = GUIGraphicsContext.Skin;
-  			}
-  			GUIGraphicsContext.Load();
-  			GUIFontManager.LoadFonts(Config.GetFile(Config.Dir.Skin, m_strSkin, "fonts.xml"));
-  			GUIFontManager.InitializeDeviceObjects();
+        GUIGraphicsContext.DX9Device.EvictManagedResources();
+        GUIWaitCursor.Dispose();
+        if (!m_strSkin.Equals(GUIGraphicsContext.Skin))
+        {
+          m_strSkin = GUIGraphicsContext.Skin;
+        }
+        GUIGraphicsContext.Load();
+        GUIFontManager.LoadFonts(Config.GetFile(Config.Dir.Skin, m_strSkin, "fonts.xml"));
+        GUIFontManager.InitializeDeviceObjects();
 
-  			if (GUIGraphicsContext.DX9Device != null)
-  			{
-  				GUIWindowManager.OnResize();
-  				GUIWindowManager.PreInit();
-  				GUIWindowManager.ActivateWindow(activeWin);
-  				GUIWindowManager.OnDeviceRestored();
-  			}
-  			// Must set the FVF after reset
-  			GUIFontManager.SetDevice();
+        if (GUIGraphicsContext.DX9Device != null)
+        {
+          GUIWindowManager.OnResize();
+          GUIWindowManager.PreInit();
+          GUIWindowManager.ActivateWindow(activeWin);
+          GUIWindowManager.OnDeviceRestored();
+        }
+        // Must set the FVF after reset
+        GUIFontManager.SetDevice();
 
-  			Log.Info("Main: Resetting DX9 device done");
-  		}
-  	}
+        Log.Info("Main: Resetting DX9 device done");
+      }
+    }
   }
 
-	#endregion
+  #endregion
 
-	/// <summary>
-	/// Handle OnResizeEnd
-	/// </summary>
-	protected override void OnResizeEnd(EventArgs e)
-	{
-	  Log.Info("Main: OnResizeEndn called");
-	  if (GUIGraphicsContext.IsDirectX9ExUsed())
-	  {
-	    _resizeOngoing = false;
-	    if (_clientSize != ClientSize)
-	    {
-	      _resetOnResize = true;
-	    }
-	  }
-	  base.OnResizeEnd(e);
-	}
+  /// <summary>
+  /// Handle OnResizeEnd
+  /// </summary>
+  protected override void OnResizeEnd(EventArgs e)
+  {
+    Log.Info("Main: OnResizeEndn called");
+    if (GUIGraphicsContext.IsDirectX9ExUsed())
+    {
+      _resizeOngoing = false;
+      if (_clientSize != ClientSize)
+      {
+        _resetOnResize = true;
+      }
+    }
+    base.OnResizeEnd(e);
+  }
 
   #region Render()
 
@@ -2235,12 +2235,12 @@ public class MediaPortalApp : D3DApp, IRender
             return;
           }
           break;
-          
+
         //toggle between directx windowed and exclusive mode
         case Action.ActionType.ACTION_TOGGLE_WINDOWED_FULLSCREEN:
           ToggleFullWindowed();
           return;
-          //break;
+        //break;
 
         //mute or unmute audio
         case Action.ActionType.ACTION_VOLUME_MUTE:
@@ -2292,28 +2292,28 @@ public class MediaPortalApp : D3DApp, IRender
         //switch between several home windows
         case Action.ActionType.ACTION_SWITCH_HOME:
           GUIMessage homeMsg;
-          if (GUIWindowManager.ActiveWindow == (int) GUIWindow.Window.WINDOW_HOME)
+          if (GUIWindowManager.ActiveWindow == (int)GUIWindow.Window.WINDOW_HOME)
           {
             homeMsg =
               new GUIMessage(GUIMessage.MessageType.GUI_MSG_GOTO_WINDOW, 0, 0, 0,
-                             (int) GUIWindow.Window.WINDOW_SECOND_HOME, 0, null);
+                             (int)GUIWindow.Window.WINDOW_SECOND_HOME, 0, null);
           }
-          else if (GUIWindowManager.ActiveWindow == (int) GUIWindow.Window.WINDOW_SECOND_HOME)
+          else if (GUIWindowManager.ActiveWindow == (int)GUIWindow.Window.WINDOW_SECOND_HOME)
           {
             homeMsg = new GUIMessage(GUIMessage.MessageType.GUI_MSG_GOTO_WINDOW, 0, 0, 0,
-                                     (int) GUIWindow.Window.WINDOW_HOME, 0, null);
+                                     (int)GUIWindow.Window.WINDOW_HOME, 0, null);
           }
           else
           {
             if (_startWithBasicHome)
             {
               homeMsg = new GUIMessage(GUIMessage.MessageType.GUI_MSG_GOTO_WINDOW, 0, 0, 0,
-                                       (int) GUIWindow.Window.WINDOW_SECOND_HOME, 0, null);
+                                       (int)GUIWindow.Window.WINDOW_SECOND_HOME, 0, null);
             }
             else
             {
               homeMsg = new GUIMessage(GUIMessage.MessageType.GUI_MSG_GOTO_WINDOW, 0, 0, 0,
-                                       (int) GUIWindow.Window.WINDOW_HOME, 0, null);
+                                       (int)GUIWindow.Window.WINDOW_HOME, 0, null);
             }
           }
           GUIWindowManager.SendThreadMessage(homeMsg);
@@ -2440,7 +2440,7 @@ public class MediaPortalApp : D3DApp, IRender
         case Action.ActionType.ACTION_SHUTDOWN:
           {
             Log.Info("Main: Shutdown dialog");
-            GUIDialogMenu dlg = (GUIDialogMenu) GUIWindowManager.GetWindow((int) GUIWindow.Window.WINDOW_DIALOG_MENU);
+            GUIDialogMenu dlg = (GUIDialogMenu)GUIWindowManager.GetWindow((int)GUIWindow.Window.WINDOW_DIALOG_MENU);
             if (dlg != null)
             {
               dlg.Reset();
@@ -2453,7 +2453,7 @@ public class MediaPortalApp : D3DApp, IRender
               //RestartOptions option = RestartOptions.Suspend;
               if (dlg.SelectedId < 0)
               {
-                GUIWindow win = GUIWindowManager.GetWindow((int) GUIWindow.Window.WINDOW_HOME);
+                GUIWindow win = GUIWindowManager.GetWindow((int)GUIWindow.Window.WINDOW_HOME);
                 if (win != null)
                 {
                   win.OnAction(new Action(Action.ActionType.ACTION_MOVE_LEFT, 0, 0));
@@ -2606,7 +2606,7 @@ public class MediaPortalApp : D3DApp, IRender
             if (!g_Player.IsTV || !GUIGraphicsContext.IsFullScreenVideo)
             {
               // Don't start playing from the beginning if we press play to return to normal speed
-              if (g_Player.IsMusic && g_Player.Speed != 1) 
+              if (g_Player.IsMusic && g_Player.Speed != 1)
               {
                 // Attention: GUIMusicGenre / GUIMusicFiles need to be handled differently. we reset the speed there
                 if (GUIWindowManager.ActiveWindow == (int)GUIWindow.Window.WINDOW_MUSIC_FILES ||
